+++ _test/var14.go
@@ -1,22 +1,9 @@
 package main
 
-var (
-	a = concat("hello", b)
-	b = concat(" ", c, "!")
-	c = d
-	d = "world"
-)
-
-func concat(a ...string) string {
-	var s string
-	for _, ss := range a {
-		s += ss
-	}
-	return s
-}
+import "github.com/containous/yaegi/_test/vars"
 
 func main() {
-	println(a)
+	println(vars.A)
 }
 
 // Output:
