+++ _test/vars/second.go
@@ -1,1 +1,6 @@
 package vars
+
+var (
+	B = concat(" ", C, "!")
+	D = "world"
+)
